% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/giorno_palette.R
\name{giorno_palette}
\alias{giorno_palette}
\title{giorno palette}
\usage{
giorno_palette(n, name = "anime", type = c("discrete", "continuous"))
}
\arguments{
\item{n}{number of colors}

\item{name}{anime or manga}

\item{type}{discrete or continuous}
}
\value{
an object of class palette
}
\description{
This function makes a palette based on the colors of giorno Cujoh
}
