% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jolyne_palette.R
\name{jolyne_palette}
\alias{jolyne_palette}
\title{jolyne palette}
\usage{
jolyne_palette(n, name = "anime", type = c("discrete", "continuous"))
}
\arguments{
\item{n}{number of colors}

\item{name}{anime or manga}

\item{type}{discrete or continuous}
}
\value{
an object of class palette
}
\description{
This function makes a palette based on the colors of Jolyne Cujoh
}
