% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joseph_palette.R
\name{joseph_palette}
\alias{joseph_palette}
\title{joseph palette}
\usage{
joseph_palette(n, name = "anime", type = c("discrete", "continuous"))
}
\arguments{
\item{n}{number of colors}

\item{name}{anime or manga}

\item{type}{discrete or continuous}
}
\value{
an object of class palette
}
\description{
This function makes a palette based on the colors of joseph Cujoh
}
