% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jotaro_palette.R
\name{jotaro_palette}
\alias{jotaro_palette}
\title{jotaro palette}
\usage{
jotaro_palette(n, name = "anime", type = c("discrete", "continuous"))
}
\arguments{
\item{n}{number of colors}

\item{name}{anime or manga}

\item{type}{discrete or continuous}
}
\value{
an object of class palette
}
\description{
This function makes a palette based on the colors of jotaro Cujoh
}
