% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/giorno_palette.R
\name{scale_color_giorno}
\alias{scale_color_giorno}
\title{scale color giorno}
\usage{
scale_color_giorno(name = "anime", type = "continuous")
}
\arguments{
\item{name}{anime or manga}

\item{type}{discrete or continuous}
}
\value{
a scale color function
}
\description{
This function specifies color using the giorno palette
}
