% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jolyne_palette.R
\name{scale_color_jolyne}
\alias{scale_color_jolyne}
\title{scale color jolyne}
\usage{
scale_color_jolyne(name = "anime", type = "continuous")
}
\arguments{
\item{name}{anime or manga}

\item{type}{discrete or continuous}
}
\value{
a scale color function
}
\description{
This function specifies color using the jolyne palette
}
