% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jonathan_palette.R
\name{scale_color_jonathan}
\alias{scale_color_jonathan}
\title{scale color jonathan}
\usage{
scale_color_jonathan(name = "anime", type = "continuous")
}
\arguments{
\item{name}{anime or manga}

\item{type}{discrete or continuous}
}
\value{
a scale color function
}
\description{
This function specifies color using the jonathan palette
}
