% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joseph_palette.R
\name{scale_color_joseph}
\alias{scale_color_joseph}
\title{scale color joseph}
\usage{
scale_color_joseph(name = "anime", type = "continuous")
}
\arguments{
\item{name}{anime or manga}

\item{type}{discrete or continuous}
}
\value{
a scale color function
}
\description{
This function specifies color using the joseph palette
}
