% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/josuke_palette.R
\name{scale_color_josuke}
\alias{scale_color_josuke}
\title{scale color josuke}
\usage{
scale_color_josuke(name = "anime", type = "continuous")
}
\arguments{
\item{name}{anime or manga}

\item{type}{discrete or continuous}
}
\value{
a scale color function
}
\description{
This function specifies color using the josuke palette
}
