% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jotaro_palette.R
\name{scale_color_jotaro}
\alias{scale_color_jotaro}
\title{scale color jotaro}
\usage{
scale_color_jotaro(name = "anime", type = "continuous")
}
\arguments{
\item{name}{anime or manga}

\item{type}{discrete or continuous}
}
\value{
a scale color function
}
\description{
This function specifies color using the jotaro palette
}
