% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/giorno_palette.R
\name{scale_fill_giorno}
\alias{scale_fill_giorno}
\title{scale fill giorno}
\usage{
scale_fill_giorno(name = "anime", type = "continuous")
}
\arguments{
\item{name}{anime or manga}

\item{type}{discrete or continuous}
}
\value{
a fill color function
}
\description{
This function specifies fill using the giorno palette
}
