% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jolyne_palette.R
\name{scale_fill_jolyne}
\alias{scale_fill_jolyne}
\title{scale fill jolyne}
\usage{
scale_fill_jolyne(type = "continuous")
}
\arguments{
\item{type}{discrete or continuous}
}
\value{
a fill color function
}
\description{
This function specifies fill using the jolyne palette
}
