% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jonathan_palette.R
\name{scale_fill_jonathan}
\alias{scale_fill_jonathan}
\title{scale fill jonathan}
\usage{
scale_fill_jonathan(name = "anime", type = "continuous")
}
\arguments{
\item{name}{anime or manga}

\item{type}{discrete or continuous}
}
\value{
a fill color function
}
\description{
This function specifies fill using the jonathan palette
}
