% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joseph_palette.R
\name{scale_fill_joseph}
\alias{scale_fill_joseph}
\title{scale fill joseph}
\usage{
scale_fill_joseph(name = "anime", type = "continuous")
}
\arguments{
\item{name}{anime or manga}

\item{type}{discrete or continuous}
}
\value{
a fill color function
}
\description{
This function specifies fill using the joseph palette
}
