% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/josuke_palette.R
\name{scale_fill_josuke}
\alias{scale_fill_josuke}
\title{scale fill josuke}
\usage{
scale_fill_josuke(name = "anime", type = "continuous")
}
\arguments{
\item{name}{anime or manga}

\item{type}{discrete or continuous}
}
\value{
a fill color function
}
\description{
This function specifies fill using the josuke palette
}
