% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jotaro_palette.R
\name{scale_fill_jotaro}
\alias{scale_fill_jotaro}
\title{scale fill jotaro}
\usage{
scale_fill_jotaro(name = "anime", type = "continuous")
}
\arguments{
\item{name}{anime or manga}

\item{type}{discrete or continuous}
}
\value{
a fill color function
}
\description{
This function specifies fill using the jotaro palette
}
